Add PermutableArray class
It's only working with values, at the moment (permuting fields in a
block, for example, does not work).

To the point of this branch, prefill a playboard by permutations, then
randomize by rotating blocks, and validate the playboard on the
browser to observe the symmetries. There's more to do in this direction.